Moved all fixtures into conftest
